move analysis implementation to the internal directory
